requirements.txt
scikit_image==0.14.2
imutils==0.5.2
tensorflow==1.12.0
numpy==1.15.4
watchdog==0.9.0
Keras==2.2.4
Pillow==5.4.1
skimage==0.0
scikit_learn==0.20.2
